+++ src/main/java/org/example/components/Book.java
@@ -1,0 +1,37 @@
+package org.example.components;
+
+public class Book extends Publication implements Printable {
+    private String ISBN;
+
+    public Book(String title, String author, int year, String ISBN) {
+        super(title, author, year);
+        this.ISBN = ISBN;
+    }
+
+    public String getISBN() {
+        return ISBN;
+    }
+
+    public void setISBN(String ISBN) {
+        this.ISBN = ISBN;
+    }
+
+    @Override
+    public String getType() {
+        return "Book";
+    }
+
+    @Override
+    public String toString() {
+        return super.toString() + ", ISBN='" + ISBN + "'}";
+    }
+
+    @Override
+    public void printDetails() {
+        System.out.println("Type: Book");
+        System.out.println("Name: " + getTitle());
+        System.out.println("Author: " + getAuthor());
+        System.out.println("Year: " + getYear());
+        System.out.println("ISBN: " + ISBN);
+    }
+}
