+++ src/main/java/org/example/components/Printable.java
@@ -1,0 +1,5 @@
+package org.example.components;
+
+public interface Printable {
+    void printDetails();
+}
